requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.3
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
stripe==2.60.0
urllib3==1.26.6
django_heroku
gunicorn
